# perf: Gross Profit report will fetch SLE's on demand and memoize