Stock Ledger Entries are fetched on demand and memoized in dictionary. Testing on ~1.7 million SLE's
```
In [4]: frappe.db.sql("select count(name) from `tabStock Ledger Entry` where is_cancelled=0")
Out[4]: ((1716875,),)
```

Filters:
```
In [2]: f={"company":"testsite","from_date":"2022-11-01","to_date":"2023-01-20","group_by":"Invoice"}
   ...: gp=frappe.get_doc("Report","Gross Profit")
   ...: 
```

## Original
### Memory usage
```
In [3]: %memit res=gp.execute_script_report(filters=f)
   ...: 
peak memory: 1708.94 MiB, increment: 1605.23 MiB

In [4]: len(res[1])
Out[4]: 12173
```
### Runtime
```
In [2]: %timeit -n 1 -r 1 res=gp.execute_script_report(filters=f)
2min 24s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```

## Refactored
### Memory usage
```
In [4]: %memit res=gp.execute_script_report(filters=f)
peak memory: 430.59 MiB, increment: 328.48 MiB

In [6]: len(res[1])
Out[6]: 12173
```

### Runtime
```
In [2]: %timeit -n 1 -r 1 res=gp.execute_script_report(filters=f)
50.5 s ± 0 ns per loop (mean ± std. dev. of 1 run, 1 loop each)
```
This reduction in runtime `2min 24s` -> `50.5s` and memory usage `1708.94 MiB` -> `430.59 MiB` is purely from not fetching unwanted stock ledger entries.

This is just a temporary fix. With a sufficiently high cardinality for `item_code` and `warehouse` in the filtered date range, this approach will fetch most of the stock ledger similar to the original approach. Will have to figure out a more stable solution, which probably uses windows functions to calculate stock_value_change.