@rtdany10 2023-02-18
@ruthra-kumar maybe we can also remove the unnecessary inner join while fetching incoming rate from DN?